+++ irdap/config.conf
@@ -20,6 +20,8 @@ normalized_polarization_images  = False
 [Basic ADI options]
 
 perform_adi                     = False
+principal_components		= disk
+pca_radii			= [20,100,512]
 
 [Advanced pre-processing options]
 # Generally do not need to be changed
